Tested with placeholder JSON string.

+++ app/src/main/java/com/example/android/quakereport/Quake.java
@@ -4,18 +4,18 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 
 public class Quake {
-    private float magnitude;
+    private double magnitude;
     private String location;
     private Date date;
 
-    public Quake(float magnitude, String location, Long time) {
+    public Quake(double magnitude, String location, Long time) {
         this.magnitude = magnitude;
         this.location = location;
         this.date=new Date();
         date.setTime(time);
     }
 
-    public float getMagnitude() {
+    public double getMagnitude() {
         return magnitude;
     }
